refactor(web): atomize pipeline routes
